Refactor record encryption to EncryptedRecord
