+++ dpc/src/testnet2/record/tests.rs
@@ -15,7 +15,7 @@
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    testnet2::{encoded::*, instantiated::*, payload::Payload, record_encryption::*, DPC},
+    testnet2::{encoded::*, encrypted::*, instantiated::*, payload::Payload, DPC},
     traits::{AccountScheme, DPCComponents, EncodedRecordScheme},
     Account,
     ViewKey,
@@ -148,8 +148,7 @@ fn test_record_encryption() {
             .unwrap();
 
             // Encrypt the record
-            let (_, encryped_record) =
-                RecordEncryption::encrypt_record(&system_parameters, &given_record, &mut rng).unwrap();
+            let (_, encryped_record) = EncryptedRecord::encrypt(&system_parameters, &given_record, &mut rng).unwrap();
             let account_view_key = ViewKey::from_private_key(
                 &system_parameters.account_signature,
                 &system_parameters.account_commitment,
@@ -159,7 +158,7 @@ fn test_record_encryption() {
 
             // Decrypt the record
             let decrypted_record =
-                RecordEncryption::decrypt_record(&system_parameters, &account_view_key, &encryped_record).unwrap();
+                EncryptedRecord::decrypt(&system_parameters, &account_view_key, &encryped_record).unwrap();
 
             assert_eq!(given_record, decrypted_record);
         }
